requirements.txt
llama-index-storage-docstore-mongodb
llama-index-storage-index-store-mongodb
llama-index-llms-openai
llama-index

llama-index-storage-docstore-redis
llama-index-storage-index-store-redis
llama-index-llms-langchain

llama-index-embeddings-huggingface
llama-index-embeddings-langchain

llama-index-vector-stores-mongodb

transformers
# Additional runtime dependencies (ensure compatible versions in CI)
python-dotenv>=0.21.0
openai>=0.27.0
SQLAlchemy>=1.4,<2.2
chromadb>=0.3.28
redis>=4.5.0
pymongo>=4.2.0
requests>=2.28.0
crawl4ai>=0.2.0
llama-index>=0.10.0,<0.13.0
